add a top-level !user command with subcommands
This replaces the old !user command (which displayed user info) with a
new top level !user command with the following sub-commands (localtime,
hello, info).

Had to implement aliases for these 3 subcommands manually, as maubot
wont let you alias a command-level to a command + subcommand. So these
are done with a passive command.

Signed-off-by: Ryan Lerch <rlerch@redhat.com>