Fix User Timezone Null issue
Fixes Null Timezone Error: Previously, attempting to add a user to the on-call list would silently fail (on the Matrix side) if their timezone wasn't defined in Fedora Accounts. This PR introduces a fallback mechanism that gracefully handles users without a set timezone by using "UTC" as the default.

Issue: can't add a user to the oncall list if they havent defined a timezone #71
Description: The code attempted to insert the user's timezone into the database, but it raised a NotNullViolationError if the user hadn't set a timezone.
Solution: The code now retrieves the user's timezone and defaults to "UTC" if it's None. This ensures users can be added to the on-call list regardless of their timezone preference.

Signed-off-by: [KimFarida] <[farimomoh@gmail.com]>